feat: webpack-dev-server base setup done
